Refactored ReOrganise and Shortcut layer
