+++ keras_cv/models/object_detection/yolov7/__internal__/layers/yolov7_encoder.py
@@ -11,14 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-import keras_cv.models.object_detection.yolov7 as yolov7
-from yolov7.__internal__.layers.helpers import (
-    ReOrg,
-    Shortcut,
-    FusedConvolution,
-    DownC,
-    ImplicitAddition,
-    ImplicitMultiplication,
-    SPPCSPC,
-)
